pyproject.toml
[project]
name = "play-eclipse-ditto"
version = "0.2.0"
description = "Examples to understand and test eclipse-ditto"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "python-dotenv>=1.1.1",
    "colorama>=0.4.6",
    "httpx>=0.28.1",
    "aiohttp>=3.9.0",
    "pandas>=2.3.1",
    "tqdm>=4.67.1",
    "poethepoet>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = [
    "black>=24.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
    "hurl>=1.10.0",
]



[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.ruff.lint]
select = ["E", "F", "W", "B", "Q", "I", "ASYNC", "T20"]
ignore = ["F401", "E501", "T201"]  # Ignore print statements as they're intentional for user feedback

[tool.ruff.lint.flake8-tidy-imports]
[tool.ruff.lint.flake8-tidy-imports.banned-api]
"unittest".msg = "Use `pytest` instead."

[tool.poe]
include = "poe.toml"
